opcode: Fix read_opc_u4() return type

+++ include/hornet/opcode.hh
@@ -20,7 +20,7 @@ inline uint16_t read_opc_u2(char *p)
     return read_opc_u1(p) << 8 | read_opc_u1(p+1);
 }
 
-inline uint16_t read_opc_u4(char *p)
+inline uint32_t read_opc_u4(char *p)
 {
     return read_opc_u1(p) << 24 | read_opc_u1(p+1) << 16 | read_opc_u1(p+2) << 8 | read_opc_u1(p+3);
 }
